requirements.txt
python-dotenv
pandas
cassandra-driver
boto3